+++ packages/database/src/repositories/index.ts
@@ -2,4 +2,5 @@ export * from "./custom-settings.repository";
 export * from "./mod.repository";
 export * from "./mod-download.repository";
 export * from "./profile.repository";
+export * from "./rss-item.repository";
 export * from "./vpk.repository";