+++ packages/database/src/schema/index.ts
@@ -3,4 +3,5 @@ export * from "./custom-settings";
 export * from "./job-locks";
 export * from "./mods";
 export * from "./profiles";
+export * from "./rss-items";
 export * from "./vpk";